+++ risk/margins_calculation.go
@@ -107,8 +107,12 @@ func (e *Engine) calculateMargins(m events.Margin, markPrice *num.Uint, rf types
 			if auction {
 				exitPrice = e.ob.GetIndicativePrice()
 			} else {
-				svol, _ := slippageVolume.Float64()
-				exitPrice, err = e.ob.GetCloseoutPrice(uint64(svol), types.SideBuy)
+				svol := slippageVolume.Abs()
+				if !svol.IsInteger() {
+					e.log.Debug("slippage volume should be decimal but isn't",
+						logging.Error(err))
+				}
+				exitPrice, err = e.ob.GetCloseoutPrice(uint64(svol.IntPart()), types.SideBuy)
 				if err != nil && e.log.GetLevel() == logging.DebugLevel {
 					e.log.Debug("got non critical error from GetCloseoutPrice for Buy side",
 						logging.Error(err))
@@ -148,8 +152,12 @@ func (e *Engine) calculateMargins(m events.Margin, markPrice *num.Uint, rf types
 			if auction {
 				exitPrice = e.ob.GetIndicativePrice()
 			} else {
-				svol, _ := slippageVolume.Abs().Float64()
-				exitPrice, err = e.ob.GetCloseoutPrice(uint64(svol), types.SideSell)
+				svol := slippageVolume.Abs()
+				if !svol.IsInteger() {
+					e.log.Debug("slippage volume should be decimal but isn't",
+						logging.Error(err))
+				}
+				exitPrice, err = e.ob.GetCloseoutPrice(uint64(svol.IntPart()), types.SideSell)
 				if err != nil && e.log.GetLevel() == logging.DebugLevel {
 					e.log.Debug("got non critical error from GetCloseoutPrice for Sell side",
 						logging.Error(err))